check if load-bety.sh already exists before downloading.

+++ script/load.bety.sh
@@ -3,6 +3,8 @@
 # Wrapper for script used to load and sync BETYdb servers
 # Documentation: https://github.com/PecanProject/pecan/wiki/Database-Synchronization
 
-wget https://raw.githubusercontent.com/PecanProject/pecan/master/scripts/load.bety.sh -o load.bety.sh
-chmod +x load.bety.sh
+if [ ! -e load.bety.sh ];
+  wget https://raw.githubusercontent.com/PecanProject/pecan/master/scripts/load.bety.sh -o load.bety.sh
+  chmod +x load.bety.sh
+fi
 ./load.bety.sh $*
